# executor: optimize count distinct with single column