<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->

Resolves #7539

### What is changed and how it works?

Treat individually if `count distinct` has only 1 argument.

```
Count/count(distinct)/bigint(20)-12                     3.68ms ± 1%  1.86ms ± 1%  -49.49%  (p=0.008 n=5+5)
Count/count(distinct)/float-12                          3.62ms ± 2%  2.35ms ± 2%  -35.07%  (p=0.008 n=5+5)
Count/count(distinct)/double-12                         3.59ms ± 2%  2.34ms ± 2%  -34.72%  (p=0.008 n=5+5)
Count/count(distinct)/decimal(11,0)-12                  6.72ms ± 1%  5.49ms ± 1%  -18.28%  (p=0.008 n=5+5)
Count/count(distinct)/char(1)-12                        4.62ms ± 3%  2.56ms ± 4%  -44.66%  (p=0.008 n=5+5)
Count/count(distinct)/date-12                           4.18ms ± 1%  4.35ms ± 6%   +3.99%  (p=0.008 n=5+5)
Count/count(distinct)/time-12                           3.85ms ± 2%  1.80ms ± 4%  -53.37%  (p=0.008 n=5+5)
Count/count(distinct)/json-12                           4.05ms ± 1%  4.04ms ± 2%     ~     (p=0.841 n=5+5)
```


Before,
```mysql
MySQL [test]> select count(distinct id) from test;
+--------------------+
| count(distinct id) |
+--------------------+
|           10000000 |
+--------------------+
1 row in set (17.64 sec)
```

After,
```mysql
MySQL [test]> select count(distinct id) from test;
+--------------------+
| count(distinct id) |
+--------------------+
|           10000000 |
+--------------------+
1 row in set (10.85 sec)
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test


Side effects

 - Increased code complexity